create bundle directory earlier
Before this, container bundle directory was initialized and created in
spec handler, but it was used by storage handler and io handler which
before spec handler, bundle directory was still "" in storage handler
and io handler, which caused files were created under "/".
So we should initialize and create bundle directory in metadata add
handler which is before all other ones.

Signed-off-by: Vanient <xiadanni1@huawei.com>