wasm: allow BSD-2-Clause
The crate of "mach" which is indirectly dependence introduces by
wasmtime-runtime, is in the license of BSD-2-Clause, we have to allow
it, otherwise we can not use wasmtime.

The Modified or New BSD (or BSD 3-clause) license is the same as BSD-2,
but with an additional clause prohibiting the names of the authors from
being used to endorse or promote products relating to the software.

As BSD-3-Clause is allowed, we have to reason to deny BSD-2-Clause