clear pod cpuset
When the kubelet configures cpuManagerPolicy as static, the Pod will specify CPU IDs for CPU affinity. Due to the different cpusets of the guest OS and host OS, this can lead to configuration failures. There is no need for CPU affinity in the guest OS, so the Pod's cpuset should be cleared.

Signed-off-by: MorningTZH <morningtzh@yeah.net>