vmm: fix stdout/stderr preemption
We introduced a preemptable receiver feature so that the new start containerd
can preempt the old stdio streaming process routine.

but it has a bug that when io copy thread finished, it will remove the
io channel, which has a preemption signal sender in it, the removal of
the sender may preempt the stream receiver to stop early, without all
message processed.

After this PR change, after copy of stdout and stderr finished, we don't
call remove of the iochannel, but just set a flag of sender_closed to
true. Before stream handle return, it just determin if the io channel
should be removed by this flag.

Signed-off-by: Abel Feng <fshb1988@gmail.com>